Coding style fixes in src/hl.c and src/state.c

+++ src/hl.h
@@ -4,7 +4,7 @@
 #include "color.h"
 #include "buffer.h"
 
-HlColor **hl_line(Buffer *b, const char *line, int len, int line_nr, int *next_changed);
+HlColor **hl_line(Buffer *b, const char *line, size_t len, int line_nr, int *next_changed);
 void hl_fill_start_states(Buffer *b, int line_nr);
 void hl_insert(Buffer *b, int first, int lines);
 void hl_delete(Buffer *b, int first, int lines);
